requirements.txt
google-genai>=0.3.0
google-generativeai>=0.8.0
google-adk>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0